modify: PrimalityTestingのクラス階層
