+++ test/matsu/num/number/primes/MillerPrimalityIntTest.java
@@ -21,7 +21,15 @@ final class MillerPrimalityIntTest {
 
     private static final PrimalityInt Test_Primality = new MillerPrimalityInt();
 
-    public static class Enumeration extends PrimalityIntTesting.Enumeration {
+    public static class Enumeration extends PrimalityTesting.EnumerationInt {
+
+        @Override
+        PrimalityInt getPrimalityInt() {
+            return Test_Primality;
+        }
+    }
+
+    public static class Randoms extends PrimalityTesting.RandomsInt {
 
         @Override
         PrimalityInt getPrimalityInt() {
