+++ test/matsu/num/number/primes/MillerPrimalityLongTest.java
@@ -21,7 +21,7 @@ final class MillerPrimalityLongTest {
 
     private static final PrimalityLong Test_Primality = new MillerPrimalityLong();
 
-    public static class Enumeration extends PrimalityLongTesting.Enumeration {
+    public static class Enumeration extends PrimalityTesting.EnumerationLong {
 
         @Override
         PrimalityLong getPrimalityLong() {
@@ -29,7 +29,7 @@ final class MillerPrimalityLongTest {
         }
     }
 
-    public static class Randoms extends PrimalityLongTesting.Randoms {
+    public static class Randoms extends PrimalityTesting.RandomsLong {
 
         @Override
         PrimalityLong getPrimalityLong() {
